+++ src/components/Charts.tsx
@@ -1,6 +1,6 @@
 import { useMemo } from "react";
 import { Bar, BarChart, Cell, Legend, Pie, PieChart, ResponsiveContainer, Tooltip } from "recharts";
-import { t } from "../i18n-utils";
+import { t } from "../util/i18n";
 
 import { COLORS } from "../constants";
 import type { ChartData } from "../schemas/statitics";
