+++ src/components/DataAvailabilityForm.tsx
@@ -4,7 +4,7 @@ import { useState } from "react";
 import { FormProvider, type SubmitHandler, useForm } from "react-hook-form";
 import { useTranslation as useT } from "react-i18next";
 import { type TypeOf, object, string } from "zod";
-import { t as tNative } from "../i18n-utils";
+import { t as tNative } from "../util/i18n";
 
 import FormInput from "../stories/form-input/FormInput";
 import LinkButton from "../stories/link-button/LinkButton";
